+++ backend/config/db.js
@@ -1,0 +1,5 @@
+const mongoose = require("mongoose");
+
+mongoose.connect("mongodb+srv://sanda:TVRS1234%23@cluster0.r6puny8.mongodb.net/?retryWrites=true&w=majority&appName=Cluster0");
+
+module.exports = mongoose;
