+++ backend/middlewares/uploadMiddleware.js
@@ -1,0 +1,17 @@
+const jwt = require("jsonwebtoken");
+
+const fetchUser = async (req, res, next) => {
+    const token = req.header('auth-token');
+    if (!token) {
+        return res.status(401).send({ errors: "Please authenticate using valid token" });
+    }
+    try {
+        const data = jwt.verify(token, 'secret_cake');
+        req.user = data.user;
+        next();
+    } catch (error) {
+        res.status(401).send({ errors: "Invalid token. Please authenticate again." });
+    }
+};
+
+module.exports = fetchUser;
