+++ backend/models/Banner.js
@@ -1,0 +1,10 @@
+const mongoose = require("mongoose");
+
+const bannerSchema = new mongoose.Schema({
+    url: { type: String, required: true },
+    date: { type: Date, default: Date.now },
+});
+
+const Banner = mongoose.model("Banner", bannerSchema);
+
+module.exports = Banner;
